REFACTOR: all supporting deploy/build functions are in utils.js now, config.js has configs.
We've also moved error codes to errors.js.
There is some tension between whether to make it code-oriented or
task-oriented.
buildCode in utils.js is a good example, it uses gulp-git but
I've put it for the moment in util.js which makes it appear like a
standalone utility that could be run outside of gulp.

We should probably press on at this point as we've done a lot of tidying
up after getting git and docker tasks to work (as part of building the
walking skeleton).